- método "findAll" de la entidad "Plato" funciona correctamente. Entidad "Plato" FINALIZADA.
- Se arregló el método "remove" de la entidad "Ingrediente", la cual no eliminaba las instancias de "IngredienteDeProveedor" asociadas cuando se intentaba eliminar el ingrediente, causando un error.

- Se arregló el método "remove" de la entidad "Proveedor" para que valide su eliminación y la permita sólo si ningún ingrediente o pedido depende únicamente de él. Posteriormente elimina la entidad junto con todas las instancias de "IngredienteDeProveedor" y "BebidaDeProveedor".

- Se incorporó la función sanitizadora a los métodos "update" de las entidades "Ingrediente" y "Plato"

- Pensar en incorporar un atributo "estado" o "activo" en la entidad "Plato" para poder darlo de baja o no. Esto porque no debería ser posible eliminar un plato si ya fue pedido en algún momento por un cliente, ya que causaría inconsistencias en el sistema.

- Se agregó la función "isConflictError" en "errorHandler.ts" para validar si un error debe enviar el código 409.

- Validamos que, al intentar eliminar una tarjeta, se verifique que no exista ninguna tarjeta de cliente de ese tipo (Visa, Mastercard, etc)

- Se agregó la función sanitizadora para la entidad "TarjetaCliente".

- Errores "UsuarioIsNotAllowedError" y "UsuarioUnauthorizedError" creados. Fueron incorporados en la entidad "TarjetaCliente".

- POR HACER:
1: Incorporar los errores "UsuarioIsNotAllowedError" y "UsuarioUnauthorizedError" en los métodos que correspondan (como los de Ingrediente, Proveedor, Plato, etc)

2: Terminar el CU de Pedido (not easy pero necesario asap)

+++ src/plato/tipoPlato.controller.ts
@@ -3,8 +3,9 @@ import { TipoPlato } from './tipoPlato.entity.js'
 import { orm } from '../shared/db/orm.js'
 import { validarTipoPlato } from './tipoPlato.schema.js'
 import { handleErrors } from '../shared/errors/errorHandler.js'
-import { TipoPlatoNotFoundError, TipoPlatoUniqueConstraintViolation } from '../shared/errors/entityErrors/tipoPlato.errors.js'
+import { TipoPlatoAlreadyInUseError, TipoPlatoNotFoundError, TipoPlatoUniqueConstraintViolation } from '../shared/errors/entityErrors/tipoPlato.errors.js'
 import { validarFindAll } from '../shared/validarFindAll.js'
+import { Plato } from './plato.entity.js'
 
 const em = orm.em
 
@@ -64,6 +65,9 @@ async function remove (req: Request, res: Response,) {
   try {
     const numPlato = Number.parseInt(req.params.numPlato)
     const deletedTipoPlato = await em.findOneOrFail(TipoPlato, {numPlato}, {failHandler: () => {throw new TipoPlatoNotFoundError}}) 
+    if((await em.find(Plato, {tipoPlato: deletedTipoPlato})).length > 0) {
+      throw new TipoPlatoAlreadyInUseError
+    }
     await em.removeAndFlush(deletedTipoPlato)
     res.status(200).json({message: `El tipo de plato ${deletedTipoPlato.descTPlato} ha sido eliminado con éxito`, data: deletedTipoPlato})
   } catch(error: any){
