+++ src/tarjetaCliente/tarjeta.controller.ts
@@ -2,29 +2,18 @@ import { Request, Response } from 'express';
 import { Tarjeta } from './tarjeta.entity.js';
 import { orm } from '../shared/db/orm.js';
 import { validarTarjeta } from './tarjeta.schema.js';
-import z from 'zod'
-import { TarjetaUniqueConstraintViolation } from '../shared/errors/entityErrors/tarjeta.errors.js';
+import { TarjetaAlreadyInUseError, TarjetaNotFoundError, TarjetaUniqueConstraintViolation } from '../shared/errors/entityErrors/tarjeta.errors.js';
+import { validarFindAll } from '../shared/validarFindAll.js';
+import { handleErrors } from '../shared/errors/errorHandler.js';
 
 const em = orm.em;
 
 em.getRepository(Tarjeta)
 
-function handleErrors(error: any, res: Response) {
-  if (error instanceof z.ZodError) {
-    res.status(400).json({ message: JSON.parse(error.message)[0].message })
-  } else if (error.name === 'NotFoundError') {
-    res.status(404).json({message: 'La tarjeta no ha sido encontrada'})
-  } else if (error.name === 'UniqueConstraintViolationException') {
-    res.status(400).json({ message: error.sqlMessage })
-  } else {
-    res.status(500).json({ message: error.message })
-  }
-}
-
 // Obtener todas las tarjetas
 async function findAll(req: Request, res: Response) {
   try {
-    const tarjetas = await em.find(Tarjeta, {});
+    const tarjetas = validarFindAll(await em.find(Tarjeta, {}), TarjetaNotFoundError)
     res.status(200).json({ message: 'Todas las tarjetas fueron encontradas', data: tarjetas });
   } catch (error: any) {
     handleErrors(error, res)
@@ -78,7 +67,15 @@ async function update(req: Request, res: Response) {
 async function remove(req: Request, res: Response) {
   try {
     const idTarjeta = Number.parseInt(req.params.idTarjeta)
-    const tarjeta = await em.findOneOrFail(Tarjeta, { idTarjeta });
+    const tarjeta = await em.findOneOrFail(Tarjeta, { idTarjeta }, {populate: ['tarjetaClientes']}); 
+    // Usamos "populate" para obtener todas las tarjetas del cliente
+
+    // Validamos que no exista ninguna tarjetaCliente que sea de este tipo de tarjeta (Visa, Mastercard, etc)
+    if(tarjeta.tarjetaClientes.length > 0) {
+      throw new TarjetaAlreadyInUseError
+    }
+    // Validamos que no exista ninguna tarjetaCliente que sea de este tipo de tarjeta (Visa, Mastercard, etc)
+
     await em.removeAndFlush(tarjeta);
     res.status(200).json({ message: 'La tarjeta ha sido eliminada con éxito', data: tarjeta });
   } catch (error: any) {
