+++ src/tarjetaCliente/tarjetaCliente.controller.ts
@@ -1,4 +1,3 @@
-
 import { Request, Response, NextFunction } from "express"
 import { orm } from "../shared/db/orm.js"
 import { Usuario } from "../usuario/usuario.entity.js"
@@ -7,16 +6,35 @@ import { validarTarjetaCliente } from "./tarjetaCliente.schema.js"
 import { handleErrors } from "../shared/errors/errorHandler.js"
 import { validarFindAll } from "../shared/validarFindAll.js"
 import { TarjetaClienteNotFoundError, TarjetaClientePreconditionFailed, TarjetaClienteUniqueConstraintViolation } from "../shared/errors/entityErrors/tarjetaCliente.errors.js"
-import { UsuarioNotFoundError } from "../shared/errors/entityErrors/usuario.errors.js"
+import { UsuarioIsNotAllowedError, UsuarioNotFoundError, UsuarioUnauthorizedError } from "../shared/errors/entityErrors/usuario.errors.js"
 import { Tarjeta } from "./tarjeta.entity.js"
+import { TarjetaNotFoundError } from "../shared/errors/entityErrors/tarjeta.errors.js"
 
 const em = orm.em
+
+function sanitizeTarjetaCliente(req: Request, res: Response, next: NextFunction) {
+  req.body.sanitizedInput = {
+    idTarjeta: req.body.idTarjeta,
+    nroTarjeta: req.body.nroTarjeta,
+    tipoTarjeta: req.body.tipoTarjeta,
+    bancoTarjeta: req.body.bancoTarjeta,
+    titular: req.body.titular,
+    vencimiento: req.body.vencimiento,
+    codSeguridad: req.body.codSeguridad,
+    tarjeta: req.body.tarjeta,
+    cliente: req.params.id
+  }
+  next()
+}
 
 //Sería interesante manejar queryString para consultas como "findAll tarjetas de crédito/débito"
 async function findAll(req:Request,res:Response) {
   try{
+    if(!req.params.id) {
+      throw new UsuarioUnauthorizedError
+    }
     const id = Number.parseInt(req.params.id)
-    const cliente = await em.findOneOrFail(Usuario, {id})
+    const cliente = await em.findOneOrFail(Usuario, {id}, {failHandler: () => {throw new UsuarioNotFoundError}})
     const tarjetasCliente = validarFindAll(await em.find(TarjetaCliente, {cliente}, {populate: ['tarjeta', 'cliente']}), TarjetaClienteNotFoundError)
     res.status (200).json({message: 'Todos las tarjetas encontradas', data: tarjetasCliente})
   } catch (error:any){
@@ -27,6 +45,9 @@ async function findAll(req:Request,res:Response) {
 //Sería bueno manejar queryString por si se quiere buscar por número de tarjeta, por ejemplo.
 async function findOne(req:Request,res:Response) {
   try{
+    if(!req.params.id) {
+      throw new UsuarioUnauthorizedError
+    }
     const idTarjeta = Number.parseInt(req.params.idTarjeta)
     const id = Number.parseInt(req.params.id)
     const cliente = await em.findOneOrFail(Usuario, {id}, {failHandler: () => {throw new UsuarioNotFoundError}})
@@ -39,11 +60,19 @@ async function findOne(req:Request,res:Response) {
 
 async function add(req:Request,res:Response) {
   try{
-    if((await em.find(Tarjeta, {})).length === 0 || (await em.find(Usuario, {tipoUsuario: 'cliente'})).length === 0) throw new TarjetaClientePreconditionFailed
+    if((await em.find(Tarjeta, {})).length === 0 || (await em.find(Usuario, {tipoUsuario: 'cliente'})).length === 0) {
+      throw new TarjetaClientePreconditionFailed
+    } else if(!req.params.id) {
+      throw new UsuarioUnauthorizedError
+    }
     const id = Number.parseInt(req.params.id)
     const cliente = await em.findOneOrFail(Usuario, {id}, {failHandler: () => {throw new UsuarioNotFoundError}})
-    req.body.cliente = cliente
-    const tarjetaClienteValida = validarTarjetaCliente(req.body)
+    if(cliente.tipoUsuario === 'empleado') {
+      throw new UsuarioIsNotAllowedError
+    }
+    req.body.sanitizedInput.cliente = cliente
+    req.body.sanitizedInput.tarjeta = await em.findOneOrFail(Tarjeta, {idTarjeta: Number.parseInt(req.body.tarjeta)}, {failHandler: () => {throw new TarjetaNotFoundError}})
+    const tarjetaClienteValida = validarTarjetaCliente(req.body.sanitizedInput)
     const tarjetaCliente = em.create(TarjetaCliente, tarjetaClienteValida)
     await em.flush()
     res.status(201).json({message: 'Tarjeta agregada', data: tarjetaCliente})
@@ -71,10 +100,13 @@ async function update(req:Request,res:Response) {
 }*/
 
 async function remove(req:Request, res:Response) {
-    try {
+  try {
+    if(!req.params.id) {
+      throw new UsuarioUnauthorizedError
+    }
     const idTarjeta = Number.parseInt(req.params.idTarjeta)
     const id = Number.parseInt(req.params.id)
-    const cliente = await em.findOneOrFail(Usuario, {id}, {failHandler: () => {throw new UsuarioNotFoundError}})
+    const cliente = await em.findOneOrFail(Usuario, {id}, {populate: [], failHandler: () => {throw new UsuarioNotFoundError}})
     const tarjetaCliente = await em.findOneOrFail(TarjetaCliente, {idTarjeta, cliente}, {failHandler: () => {throw new TarjetaClienteNotFoundError}})
     await em.removeAndFlush(tarjetaCliente)
     res.status(200).json({message: 'El plato ha sido eliminado con éxito', data: tarjetaCliente})
@@ -83,6 +115,6 @@ async function remove(req:Request, res:Response) {
   }
 }
 
-export{ findAll, findOne, add, /*update,*/ remove }
+export{ findAll, findOne, add, sanitizeTarjetaCliente, /*update,*/ remove }
